+++ powersimdata/input/converter/reise_to_grid.py
@@ -16,8 +16,8 @@ from powersimdata.input.converter.helpers import (
 from powersimdata.network.model import ModelImmutables
 
 
-class ScenarioGrid(AbstractGrid):
-    """File reader for MAT files for scenarios which were run on the server."""
+class REISEConverter(AbstractGrid):
+    """Grid builder for MAT files generated by REISE and REISE.jl"""
 
     def __init__(self, filename):
         """Constructor.
@@ -140,7 +140,7 @@ def _get_grid_model_from_scenario_list(source):
     return slm.get_scenario(scenario_number)["grid_model"]
 
 
-class FromREISE(ScenarioGrid):
+class FromREISE(REISEConverter):
     """MATLAB file reader, for MAT files created by REISE/MATPOWER"""
 
     def build(self):
@@ -149,7 +149,7 @@ class FromREISE(ScenarioGrid):
         add_information_to_model(self)
 
 
-class FromREISEjl(ScenarioGrid):
+class FromREISEjl(REISEConverter):
     """MATLAB file reader, for MAT files created (& converted) by REISE.jl"""
 
     def build(self):
@@ -307,7 +307,8 @@ def format_gencost(data):
 def add_information_to_model(grid):
     """Makes a standard grid.
 
-    :param powersimdata.input.ScenarioGrid grid: grid with missing information.
+    :param powersimdata.input.converter.reise_to_grid.REISEConverter grid: grid
+        produced by REISE/REISE.jl engines with missing information.
     """
 
     add_interconnect_to_grid_data_frames(grid)
@@ -320,7 +321,8 @@ def add_information_to_model(grid):
 def reindex_model(grid):
     """Reindex bus id columns.
 
-    :param powersimdata.input.scenario_grid.ScenarioGrid grid: grid
+    :param powersimdata.input.converter.reise_to_grid.REISEConverter grid: grid
+        produced by REISE engine to reindex.
     """
 
     def reset_id():
