+++ powersimdata/input/grid.py
@@ -1,6 +1,6 @@
 import os
 
-from powersimdata.input.scenario_grid import FromREISE, FromREISEjl
+from powersimdata.input.converter.reise_to_grid import FromREISE, FromREISEjl
 from powersimdata.network.constants.storage import storage
 from powersimdata.network.hifld.model import HIFLD
 from powersimdata.network.usa_tamu.model import TAMU
